FunctionDefinition => Symbol, FunctionCall => Referer
MDL plan: Symbol visit to FunctionCall should be able to, via
resolving FunctionDefinition target, determine if that function
is one of the pre-defined functions (combinedError1 etc.) in MDL
and if so... which and what the arguments are. Yeah. Why has
FunctionCall no parse(node)? virtual removed from Referer for now.